+++ test/test.ui.js
@@ -28,12 +28,13 @@ test('Geocoder#inputControl', function(tt) {
 
   tt.test('input', function(t) {
     setup({
-      types: 'place'
+      types: 'place',
+      mapboxgl: mapboxgl
     });
     var inputEl = container.querySelector('.mapboxgl-ctrl-geocoder input');
     var clearEl = container.querySelector('.mapboxgl-ctrl-geocoder button');
 
-    t.plan(7);
+    t.plan(9);
 
     geocoder.on(
       'loading',
@@ -47,6 +48,7 @@ test('Geocoder#inputControl', function(tt) {
       'result',
       once(function() {
         t.ok(inputEl.value, 'value populates in input');
+        t.ok(geocoder.mapMarker, 'a marker is created to show the selection')
         clearEl.dispatchEvent(clickEvent);
       })
     );
@@ -56,6 +58,7 @@ test('Geocoder#inputControl', function(tt) {
       once(function() {
         t.pass('input was cleared');
         t.equals(geocoder.fresh, false, 'the geocoder is fresh again')
+        t.equals(geocoder.mapMarker, null, 'the marker was reset on clear')
 
         geocoder.setInput('Paris');
         t.equals(inputEl.value, 'Paris', 'value populates in input');
@@ -122,7 +125,7 @@ test('Geocoder#inputControl', function(tt) {
       'placeholder is localized based on language'
     );
     t.end();
-  })
+  });
 
   tt.test('_clear is not called on keydown (tab), no focus trap', function(t){
     t.plan(3);
@@ -204,6 +207,18 @@ test('Geocoder#inputControl', function(tt) {
     });
     var wrapper = container.querySelector('.mapboxgl-ctrl-geocoder');
     t.equal(wrapper.classList.contains('geocoder-collapsed'), false, 'mapboxgl-ctrl-geocoder does not have `geocoder-collapsed` class');
+    t.end();
+  });
+
+  tt.test('createIcon', function(t) {
+    t.plan(1);
+    setup({ });
+    var icon = geocoder.createIcon('search', '<path/>');
+    t.equal(
+      icon.outerHTML,
+      '<svg class="geocoder-icon geocoder-icon-search" viewBox="0 0 18 18" xml:space="preserve" width="18" height="18"><path></path></svg>',
+      'creates an svg given the class name and path'
+    );
     t.end();
   });